Reorganises tests. Adds retry for actors.
