+++ dapr/clients/http/dapr_invocation_http_client.py
@@ -23,6 +23,7 @@ from dapr.clients.grpc._helpers import MetadataTuple, GrpcMessage
 from dapr.clients.grpc._response import InvokeMethodResponse
 from dapr.clients.http.conf import CONTENT_TYPE_HEADER
 from dapr.clients.http.helpers import get_api_url
+from dapr.clients.retry import RetryPolicy
 from dapr.serializers import DefaultJSONSerializer
 from dapr.version import __version__
 
@@ -33,16 +34,17 @@ DAPR_USER_AGENT = f'dapr-python-sdk/{__version__}'
 class DaprInvocationHttpClient:
     """Service Invocation HTTP Client"""
 
-    def __init__(
-        self, timeout: int = 60, headers_callback: Optional[Callable[[], Dict[str, str]]] = None
-    ):
+    def __init__(self, timeout: int = 60,
+            headers_callback: Optional[Callable[[], Dict[str, str]]] = None,
+            retry_policy: Optional[RetryPolicy] = None, ):
         """Invokes Dapr's API for method invocation over HTTP.
 
         Args:
             timeout (int, optional): Timeout in seconds, defaults to 60.
             headers_callback (lambda: Dict[str, str]], optional): Generates header for each request.
+            retry_policy (RetryPolicy optional): Specifies retry behaviour
         """
-        self._client = DaprHttpClient(DefaultJSONSerializer(), timeout, headers_callback)
+        self._client = DaprHttpClient(DefaultJSONSerializer(), timeout, headers_callback, retry_policy=retry_policy)
 
     async def invoke_method_async(
         self,
