+++ daprdocs/content/en/python-sdk-docs/python-client.md
@@ -49,7 +49,7 @@ with DaprClient("mydomain:50051?tls=true") as d:
     # use the client
 ```  
 
-#### Environment variables:  
+#### Configuration options:  
 
 ##### Dapr Sidecar Endpoints
 You can use the standardised `DAPR_GRPC_ENDPOINT` environment variable to
@@ -81,6 +81,37 @@ The client will wait for the sidecar to be up and running before proceeding.
 The default timeout is 60 seconds, but it can be overridden by setting the `DAPR_HEALTH_TIMEOUT`
 environment variable.
 
+##### Retries and timeout
+
+The Dapr client can retry a request if a specific error code is received from the sidecar. This is
+configurable through the `DAPR_API_MAX_RETRIES` environment variable and is picked up automatically, 
+not requiring any code changes.
+The default value for `DAPR_API_MAX_RETRIES` is `0`, which means no retries will be made.  
+
+You can fine-tune more retry parameters by creating a `dapr.clients.retry.RetryPolicy` object and
+passing it to the DaprClient constructor:
+
+```python
+from dapr.clients.retry import RetryPolicy
+
+retry = RetryPolicy(
+    max_attempts=5, 
+    initial_backoff=1, 
+    max_backoff=20, 
+    backoff_multiplier=1.5,
+    retryable_http_status_codes=[408, 429, 500, 502, 503, 504],
+    retryable_grpc_status_codes=[StatusCode.UNAVAILABLE, StatusCode.DEADLINE_EXCEEDED, ]
+)
+
+with DaprClient(retry_policy=retry) as d:
+    ...
+```
+
+or for actors:
+```python
+factory = ActorProxyFactory(retry_policy=RetryPolicy(max_attempts=3))
+proxy = ActorProxy.create('DemoActor', ActorId('1'), DemoActorInterface, factory)
+```
 
 ## Error handling
 Initially, errors in Dapr followed the [Standard gRPC error model](https://grpc.io/docs/guides/error/#standard-error-model). However, to provide more detailed and informative error messages, in version 1.13 an enhanced error model has been introduced which aligns with the gRPC [Richer error model](https://grpc.io/docs/guides/error/#richer-error-model). In response, the Python SDK implemented `DaprGrpcError`, a custom exception class designed to improve the developer experience.
